moved create exported interface to /types folder
